git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2713 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/combigrid/operation/OperationMatrixLTwoExplicitFullGrid.cpp
@@ -68,8 +68,8 @@ void OperationMatrixLTwoExplicitFullGrid::buildMatrix(
 void OperationMatrixLTwoExplicitFullGrid::mult(sg::base::DataVector& alpha,
 		sg::base::DataVector& result) {
 
-	unsigned int nrows( m_->getNrows() );
-	unsigned int ncols( m_->getNcols() );
+	unsigned int nrows( static_cast<unsigned int>( m_->getNrows() ) );
+	unsigned int ncols( static_cast<unsigned int>( m_->getNcols() ) );
 
 	if (alpha.getSize() != ncols || result.getSize() != nrows) {
 		throw sg::base::data_exception("Dimensions do not match!");
